--- src/main/kotlin/fr/rushy/api/command/StopCommand.kt
@@ -1,38 +1,0 @@
-package fr.rushy.api.command
-
-import fr.rushy.api.permission.CustomPermission
-import kotlinx.coroutines.GlobalScope
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.launch
-import net.minestom.server.MinecraftServer
-import net.minestom.server.command.builder.Command
-import net.minestom.server.entity.Player
-import kotlin.time.Duration.Companion.seconds
-
-/**
- * Command to stop the server.
- */
-public class StopCommand : Command("stop") {
-
-    init {
-        setCondition { sender, _ ->
-            sender !is Player || sender.hasPermission(CustomPermission.STOP_SERVER)
-        }
-
-        setDefaultExecutor { _, _ ->
-            GlobalScope.launch {
-                MinecraftServer.getInstanceManager().instances.asSequence().flatMap {
-                    it.players
-                }.forEach {
-                    it.kick("Server closed")
-                }
-
-                delay(1.seconds)
-
-                MinecraftServer.stopCleanly()
-            }
-
-        }
-
-    }
-}
