--- src/main/kotlin/fr/rushy/api/utils/FileUtils.kt
@@ -1,9 +1,0 @@
-package fr.rushy.api.utils
-
-import java.io.File
-
-/**
- * Get the current directory where is executed the program.
- */
-public val workingDirectory: File
-    get() = File(System.getProperty("user.dir"))
