+++ src/main/kotlin/io/github/rushyverse/api/translation/SupportedLanguage.kt
@@ -1,0 +1,14 @@
+package io.github.rushyverse.api.translation
+
+import java.util.*
+
+
+/**
+ * List of supported locales to translate keys.
+ */
+public enum class SupportedLanguage(public val displayName: String, public val locale: Locale) {
+
+    ENGLISH("English", Locale("en", "gb")),
+    FRENCH("Français", Locale("fr", "fr"))
+
+}
