+++ src/main/kotlin/io/github/rushyverse/api/utils/FileUtils.kt
@@ -1,0 +1,9 @@
+package io.github.rushyverse.api.utils
+
+import java.io.File
+
+/**
+ * Get the current directory where is executed the program.
+ */
+public val workingDirectory: File
+    get() = File(System.getProperty("user.dir"))
